pyproject.toml
[tool.poetry]
name = "sms-simulator"
version = "0.1.0"
description = "Project to demonstrate SMS simulation in Python"
authors = ["Naga Venkatesh Gavini <nagavenkat9948@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pika = "^1.3.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
redis = "^5.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
